requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.1
asn1crypto==0.24.0
attrs==20.3.0
Brlapi==0.6.6
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distlib==0.3.0
distro-info===0.18ubuntu0.18.04.1
Django==3.1.7
docopt==0.6.2
filelock==3.0.12
Flask==1.1.2
httplib2==0.9.2
idna==2.9
importlib-metadata==3.7.3
importlib-resources==1.5.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
mock==4.0.3
mysql-connector-python==8.0.23
netifaces==0.10.4
numpy==1.19.5
oauth==1.0.1
olefile==0.45.1
openshot-qt==2.5.1.dev2
packaging==20.9
pandas==1.1.5
pexpect==4.2.1
Pillow==5.1.0
pipreqs==0.4.10
pluggy==0.13.1
protobuf==3.15.6
py==1.10.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.1.2
PyOpenGL==3.1.0
pyparsing==2.4.7
pyRFC3339==1.0
pytest==6.2.2
pytest-flask==1.2.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2021.1
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
reportlab==3.4.0
requests==2.23.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.15.0
sqlparse==0.4.1
system-service==0.3
systemd-python==234
toml==0.10.2
tqdm==4.42.1
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.3
vboxapi==1.0
vidcutter==6.0.0
virtualenv==20.0.20
wadllib==1.3.2
Werkzeug==1.0.1
xkit==0.0.0
yarg==0.1.9
youtube-dl==2018.3.14
zipp==3.4.1
zope.interface==4.3.2
